requirements.txt
yfinance
pandas
twstock
tabulate
colorama
lxml
python-dotenv
requests
pytest
pytest-mock